chore: rename constant files to kebab case
