+++ app/renderer/components/Inspector/HighlighterRects.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import {RENDER_CENTROID_AS} from '../../constants/SCREENSHOT';
+import {RENDER_CENTROID_AS} from '../../constants/screenshot';
 import {parseCoordinates} from '../../utils/other';
 import HighlighterCentroid from './HighlighterCentroid';
 import HighlighterRectForBounds from './HighlighterRectForBounds';
