+++ app/renderer/components/Inspector/Inspector.js
@@ -17,14 +17,14 @@ import {debounce} from 'lodash';
 import React, {useEffect, useRef, useState} from 'react';
 import {useNavigate} from 'react-router-dom';
 
-import {BUTTON} from '../../constants/ANTD_TYPES';
-import {WINDOW_DIMENSIONS} from '../../constants/COMMON';
+import {BUTTON} from '../../constants/antd-types';
+import {WINDOW_DIMENSIONS} from '../../constants/common';
 import {
   INSPECTOR_TABS,
   MJPEG_STREAM_CHECK_INTERVAL,
   SESSION_EXPIRY_PROMPT_TIMEOUT,
-} from '../../constants/SESSION_INSPECTOR';
-import {SCREENSHOT_INTERACTION_MODE} from '../../constants/SCREENSHOT';
+} from '../../constants/session-inspector';
+import {SCREENSHOT_INTERACTION_MODE} from '../../constants/screenshot';
 import {clipboard} from '../../polyfills';
 import Commands from './Commands';
 import GestureEditor from './GestureEditor';
