+++ app/renderer/components/Session/ServerTabLambdatest.js
@@ -1,7 +1,7 @@
 import {Col, Form, Input, Row} from 'antd';
 import React from 'react';
 
-import {INPUT} from '../../constants/ANTD_TYPES';
+import {INPUT} from '../../constants/antd-types';
 
 const lambdatestUsernamePlaceholder = (t) => {
   if (process.env.LAMBDATEST_USERNAME) {
