+++ app/renderer/components/Session/Session.js
@@ -4,9 +4,9 @@ import _ from 'lodash';
 import React, {useEffect} from 'react';
 import {useNavigate} from 'react-router-dom';
 
-import {BUTTON} from '../../constants/ANTD_TYPES';
-import {LINKS} from '../../constants/COMMON';
-import {ADD_CLOUD_PROVIDER_TAB_KEY} from '../../constants/SESSION_BUILDER';
+import {BUTTON} from '../../constants/antd-types';
+import {LINKS} from '../../constants/common';
+import {ADD_CLOUD_PROVIDER_TAB_KEY} from '../../constants/session-builder';
 import {ipcRenderer, shell} from '../../polyfills';
 import AdvancedServerParams from './AdvancedServerParams';
 import AttachToSession from './AttachToSession';
