+++ app/renderer/constants/common.js
