+++ app/renderer/constants/gestures.js
