+++ app/renderer/constants/screenshot.js
