+++ app/renderer/constants/session-builder.js
