+++ app/renderer/constants/session-info.js
