+++ app/renderer/constants/session-inspector.js
