+++ app/renderer/lib/appium-client.js
@@ -1,8 +1,8 @@
 import Bluebird from 'bluebird';
 import _ from 'lodash';
 
-import {SCREENSHOT_INTERACTION_MODE} from '../constants/SCREENSHOT';
-import {APP_MODE, NATIVE_APP} from '../constants/SESSION_INSPECTOR';
+import {SCREENSHOT_INTERACTION_MODE} from '../constants/screenshot';
+import {APP_MODE, NATIVE_APP} from '../constants/session-inspector';
 import {parseSource, setHtmlElementAttributes} from './webview-helpers';
 
 const {TAP, SWIPE, GESTURE} = SCREENSHOT_INTERACTION_MODE;
